try auto implement

+++ script-creare-tabele.sql
@@ -28,9 +28,9 @@ create table clienti(
   prenume VARCHAR2(30) NOT NULL,
   fonduri int,
   id_clienti INT NOT NULL PRIMARY KEY,
-  nr_telefon	NUMBER(10,0) NOT NULL,
+  nr_telefon	int NOT NULL,
   email VARCHAR2(40),
-  cnp NUMBER(10,0) NOT NULL,
+  cnp int) NOT NULL,
   created_at DATE,
   updated_at DATE
 );
